+++ plugin/src/test/java/io/jenkins/plugins/casc/SecretSourceTest.java
@@ -1,6 +1,16 @@
 package io.jenkins.plugins.casc;
 
+import io.jenkins.plugins.casc.impl.secrets.YamlSecretSource;
+import io.jenkins.plugins.casc.misc.ConfiguredWithCode;
+import org.junit.Assert;
 import org.junit.Test;
+
+import java.io.File;
+import java.io.IOException;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.file.Paths;
+import java.util.Optional;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
@@ -25,5 +35,12 @@ public class SecretSourceTest {
         assertFalse(SecretSource.requiresReveal("\\${foo}").isPresent());
     }
 
-
+    @Test
+    public void should_load_secrets_file() throws IOException, URISyntaxException {
+        URL resource = SecretSourceTest.class.getResource("secrets.yml");
+        File secretsFile = Paths.get(resource.toURI()).toFile();
+        System.setProperty("casc.secrets.config", secretsFile.getAbsolutePath());
+        YamlSecretSource yamlSecretSource = new YamlSecretSource();
+        Assert.assertEquals("secretvalue1", yamlSecretSource.reveal("secret1").get());
+    }
 }